further optimize unpacker

+++ src/kernels.h
@@ -43,9 +43,9 @@ static const float MASS_P = 0.13957039;
 
 using namespace adf;
 
-void unpack_and_filter(input_stream<int64> * __restrict in, output_stream<int32> * __restrict out0, output_stream<int32> * __restrict out1);
+void unpack_and_filter(input_stream<int64> * __restrict in, output_stream<int16> * __restrict out0, output_stream<int16> * __restrict out1);
 
-void isolation(input_stream<int32> * __restrict in0, input_stream<int32> * __restrict in1, output_stream<int16> * __restrict out0, output_stream<int16> * __restrict out1);
+void isolation(input_stream<int16> * __restrict in0, input_stream<int16> * __restrict in1, output_stream<int16> * __restrict out0, output_stream<int16> * __restrict out1);
 
 void combinatorial(input_stream<int16> * __restrict in0, input_stream<int16> * __restrict in1, output_stream<float> * __restrict out);
